allow using legacy 'builder' user in container
In its early versions, cqfd used a fixed 'builder' account in the
container. The container's user later evolved to use a user that mimics
the host environment.

In some scenarios, it may be desired to keep the legacy behavior. This
commit makes this possible, by using the following property in the
.cqfdrc file:

 [project]
 legacy_builder_user=true